% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{AddSuffix}
\alias{AddSuffix}
\alias{AddPreffix}
\title{Funciones para agregar pre/sufijos}
\usage{
AddSuffix(suffix = "")

AddPreffix(preffix = "")
}
\arguments{
\item{suffix, preffix}{caracter que luego va a agregarse.}
}
\description{
Funciones para agregar pre/sufijos
}
