% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{AnchorBreaksSym}
\alias{AnchorBreaksSym}
\title{AnchorBreaks que opcionalmente extiende los límites para ser simétrico al rededor del ancla}
\usage{
AnchorBreaksSym(
  anchor = 0,
  binwidth = NULL,
  exclude = NULL,
  bins = 10,
  symmetrical = FALSE
)
}
\description{
AnchorBreaks que opcionalmente extiende los límites para ser simétrico al rededor del ancla
}
