% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proyectar-cmip.R
\name{ceof_proyectado}
\alias{ceof_proyectado}
\title{Calcula serie temporal de un campo}
\usage{
ceof_proyectado(
  files,
  member,
  ..model,
  ..experiment,
  fields,
  time = c(lubridate::NA_Date_, "2014-12-31")
)
}
\description{
Calcula serie temporal de un campo
}
