% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-ceof.R
\name{compute_ceof}
\alias{compute_ceof}
\title{Computa Funciones Empíricas Ortogonales Complejas}
\usage{
compute_ceof(
  hgt,
  lon,
  lat,
  lev,
  time,
  temporal = FALSE,
  lats_eof = c(-90, -20),
  n = 1:2
)
}
\arguments{
\item{hgt}{vector de observaciones de altura geopotencial}

\item{lon, lat, lev}{vectores de ubicación de las observaciones de altura geopotencial}

\item{temporal}{valor lógico indicando si hay que computar anomalías temporales}

\item{lats_eof}{rango de latitudes a usar}

\item{n}{numérico indicando qué EOFs computar.}
}
\description{
Computa Funciones Empíricas Ortogonales Complejas
}
