% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{coord_polar2}
\alias{coord_polar2}
\title{Coordenadas polares con un sf blanco que tapa lo que sobra}
\usage{
coord_polar2(ymax = -20, ...)
}
\description{
Coordenadas polares con un sf blanco que tapa lo que sobra
}
