% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstats.R
\name{correlate}
\alias{correlate}
\title{Computar correlación}
\usage{
correlate(x, y, signif = 2, sep = " ", ...)
}
\arguments{
\item{x, y}{variables numéricas a correlacionar}

\item{signif}{cifras significativas para el texto}

\item{sep}{separador para usar entre el estimador y el intervalo de confianza.
Por defecto usa un espacio protejido.}
}
\description{
Capa alrededor de  \link{cor.test} que devuelve la correlación estimada, el
p-valor, el mínimo y máximo del intervalo de confianza y un texto listo
para reportar.
}
