% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstats.R
\name{correlate_complex}
\alias{correlate_complex}
\title{Correlaciona una variable compleja con una real}
\usage{
correlate_complex(z, x, angles = seq(0, 2 * pi, by = 0.5 * pi/180), ...)
}
\arguments{
\item{z}{vector complejo}

\item{x}{vector real}

\item{angles}{vector de ángulos para rotar}

\item{...}{parámetros enviados a \code{cor}}
}
\description{
Calcula la correlación de una variable real con múltiples rotaciones
de una variable compleja
}
