% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstats.R
\name{correlation_statistics}
\alias{correlation_statistics}
\title{Computa p-valor y demás en correlación}
\usage{
correlation_statistics(correlation, n, signif = 2)
}
\arguments{
\item{correlation}{vector con correlaciones}

\item{n}{vector con número de observaciones usadas para calcular la correlación}

\item{signif}{cifras significativas para el texto}
}
\description{
Similar a \code{\link[=correlate]{correlate()}} pero toma una correlación calculada y el número de observaciones.
}
