% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-datasets.R
\name{data_path}
\alias{data_path}
\title{Data locations}
\usage{
data_path(type = c("raw", "derived"), ...)
}
\arguments{
\item{type}{character indicating the type of data. The resulting base will be type_data}

\item{...}{characters to be concatenated with \code{\link[=file.path]{file.path()}}}
}
\description{
Functions that return the location of different datasets. They check that
relevant file exist and then return it with a class that has a neat print method.
}
