% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\alias{ERA5_TOC}
\alias{ERA5_geopotential}
\alias{ERA5_temperature}
\alias{ERA5_vorticity}
\alias{ERA5_ozone}
\alias{ERA5_BE}
\alias{ERA5_T2M}
\alias{NCEP_PSI}
\alias{CMAP}
\alias{ERSST}
\alias{ENSO}
\alias{DMI}
\alias{PSA}
\alias{SAM}
\alias{cEOF}
\title{Datasets}
\usage{
ERA5_TOC(force_download = FALSE, from_source = FALSE)

ERA5_geopotential(force_download = FALSE, from_source = FALSE)

ERA5_temperature(force_download = FALSE, from_source = FALSE)

ERA5_vorticity(force_download = FALSE, from_source = FALSE)

ERA5_ozone(force_download = FALSE, from_source = FALSE)

ERA5_BE(force_download = FALSE, from_source = FALSE)

ERA5_T2M(force_download = FALSE, from_source = FALSE)

NCEP_PSI(force_download = FALSE, from_source = FALSE)

CMAP(force_download = FALSE, from_source = FALSE)

ERSST(force_download = FALSE, from_source = FALSE)

ENSO(force_download = FALSE, from_source = FALSE)

DMI(force_download = FALSE, from_source = FALSE)

PSA(force_download = FALSE, from_source = FALSE)

SAM(force_download = FALSE, from_source = FALSE)

cEOF()
}
\arguments{
\item{from_source}{L贸gico indicando si hay que descargar los datos desde la fuente
primaria en vez del repositorio de Zenodo}
}
\description{
Chequea que los datos existan y si no existen los descarga de Zenodo, a menos que
\code{zenodo_url} sea \code{NULL} o se llame a la funci贸n con el argumento \code{from_source} \code{TRUE}.
En cuyo caso, descarga los datos de la fuente original con la funci贸n \code{source}.
La funci贸n advierte si el checksum no es igual al registrado.
}
