% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eof-asym.R
\name{eof_asym}
\alias{eof_asym}
\title{Computa EOF asimétrico}
\usage{
eof_asym(
  value,
  lon,
  lat,
  time,
  period = data.table::year(time) >= 1979 & data.table::year(time) <= 2000,
  n = 1
)
}
\description{
Computa EOF asimétrico
}
