% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{geom_qmap}
\alias{geom_qmap}
\title{Geom que dibuja un mapa usando geom_sf()}
\usage{
geom_qmap(
  transform = identity,
  crop = NULL,
  color = "gray50",
  size = 0.3,
  fill = NA,
  ...
)
}
\arguments{
\item{transform}{una función arbitraria que se aplicar a los datos.}

\item{crop}{caja de límites opcionales para recortar el mapa.}

\item{color, size, fill}{parámetros pasados a \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}.}

\item{...}{otros ajustes pasados a \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}.}
}
\description{
Usa un mapa global simplificado centrado en el Pacífico.
}
