% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{guide_colorsteps}
\alias{guide_colorsteps}
\alias{guide_colorsteps_bottom}
\title{guide_colorsteps personalizada}
\usage{
guide_colorsteps(even.steps = TRUE, show.limits = NULL, ticks = FALSE, ...)

guide_colorsteps_bottom(
  width = 15,
  height = 0.5,
  even.steps = FALSE,
  title.position = "top",
  show.limits = TRUE,
  ticks = TRUE,
  ...
)
}
\description{
guide_colorsteps personalizada
}
