% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is.enso34}
\alias{is.enso34}
\title{Testea si puntos están en la región ENSO 34}
\usage{
is.enso34(lon, lat)
}
\arguments{
\item{lon, lat}{vectores de longitud y latitud}
}
\value{
Un vector lógico para cada punto indicando si está dentro de la región de ENSO 34 o no
}
\description{
Testea si puntos están en la región ENSO 34
}
