% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-datasets.R
\name{new_dataset}
\alias{new_dataset}
\title{Crea una nueva base de datos}
\usage{
new_dataset(name, file, source, md5)
}
\arguments{
\item{name}{el nombre de la base de datos (usada para mensajes)}

\item{file}{el archivo donde se va a guardar}

\item{source}{una función que toma un nombre de archivo y descarga los datos}

\item{md5}{checksum para verificar que los datos están bien}
}
\value{
Una función que devuelve el nombre de archivo.
Chequea que los datos existan y si no existen los descarga de Zenodo, a menos que
\code{zenodo_url} sea \code{NULL} o se llame a la función con el argumento \code{from_source} \code{TRUE}.
En cuyo caso, descarga los datos de la fuente original con la función \code{source}.
La función advierte si el checksum no es igual al registrado.
}
\description{
Devuelve una función que hace todo lo necesario para descargar los datos y
devuelve el archivo para leer.
}
