% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise-coords.R
\name{normalise_coords}
\alias{normalise_coords}
\title{Cambiar nombres de columnas}
\usage{
normalise_coords(
  data,
  rules = list(lev = c("level"), lat = c("latitude"), lon = c("longitude", "long"), time
    = c("date"))
)
}
\arguments{
\item{data}{un dataframe.}

\item{rules}{una lista de reglas para renombrar. Cada elemento es un vector
de caracteres con el posible nombre de las columnas. El nombre que se le dará
es el nombre de elemento en la lista.}
}
\description{
Cambia los nombres de las columnas de un dataframe según un conjunto de reglas
}
