% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-ceof-cmip6.R
\name{optimise_rotation}
\alias{optimise_rotation}
\title{Rota un cEOF para maximizar la correlación con una referencia}
\usage{
optimise_rotation(ceof, ref)
}
\arguments{
\item{ceof}{cEOF a rotar}

\item{ref}{cEOF de referencia}
}
\value{
un cEOF
}
\description{
Rota un cEOF para maximizar la correlación con una referencia
}
