% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-ceof-cmip6.R
\name{pad_longitudes}
\alias{pad_longitudes}
\title{Agrega longitudes en los extremos}
\usage{
pad_longitudes(data)
}
\arguments{
\item{data}{los datos}
}
\description{
(Para poder interpolar despúes)
}
