% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstats.R
\name{partial_cor}
\alias{partial_cor}
\alias{semipartial_cor}
\title{(Semi) Partial correlation between x1 and y and x2 and y.}
\usage{
partial_cor(y, x1, x2)

semipartial_cor(y, x1, x2)
}
\arguments{
\item{y}{variable to which partial correlation will be computed}

\item{x1, x2}{variables that will be correlated with y}
}
\description{
(Semi) Partial correlation between x1 and y and x2 and y.
}
