% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstats.R
\name{Pvaluate}
\alias{Pvaluate}
\title{Computa p-valores}
\usage{
Pvaluate(estimate, std.error, df, adjustment = "none")
}
\arguments{
\item{estimate}{estimador}

\item{std.error}{error estándar}

\item{df}{grados de libertad}

\item{adjustment}{método para ajustar el p-valor (\code{\link[stats:p.adjust]{stats::p.adjust()}})}
}
\description{
Computa p-valores uando la distribución t-student a partir de estimación,
error estándar  y grados de libertad. Opcionalmente los ajusta usando FDR u
otro ajuste.
}
