% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cmip-raw.R
\name{read_cmip_raw}
\alias{read_cmip_raw}
\title{Leer un archivo de CMIP como un array}
\usage{
read_cmip_raw(files, member, vars = NULL, subset = NULL, stop_if_na = FALSE)
}
\description{
Leer un archivo de CMIP como un array
}
