% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReIm.R
\name{ReIm}
\alias{ReIm}
\alias{sep_ReIm}
\alias{factor_ReIm}
\title{Separa números complejos entre su parte real e imaginaria}
\usage{
ReIm(complex)

sep_ReIm(data, column = NULL, format = c("longer", "wider"))

factor_ReIm(part)
}
\arguments{
\item{complex}{vector de valores compeljos}

\item{data}{un dataframe}

\item{column}{nombre puro de la columna que hay que dividir. Si es \code{NULL},
la función busca si hay una única columna compleja.}

\item{format}{formato de salida. \code{"longer"} usa \link[tidyr:pivot_longer]{tidyr::pivot_longer} para devolver una
columna única con el valor con una fila para cada parte. \code{"wider"} deja una columna
llamara \code{Real} y otra llamada \code{Imaginario}}

\item{part}{vector de caracteres con elementos llamados "Real" e "Imaginario"}
}
\description{
Separa números complejos entre su parte real e imaginaria
}
