% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rm_intercept}
\alias{rm_intercept}
\title{Remueve el intercepto}
\usage{
rm_intercept(data)
}
\arguments{
\item{data}{un dataframe}
}
\description{
Remueve las filas donde la columna \code{term} es igual a \code{"(Intercept)"}.
}
