% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstats.R
\name{rotate_ceof}
\alias{rotate_ceof}
\title{Rotar un cEOF}
\usage{
rotate_ceof(x, variable, rot)
}
\arguments{
\item{x}{cEOF}

\item{variable}{nombre de la variable a rotar (sin comillas)}

\item{rot}{data.table que se une a x con una variable llamada
\code{angle}con el ángulo a rotar.}
}
\value{
un cEOF
}
\description{
Rotar un cEOF
}
