% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve-poisson.R
\name{solve_poisson}
\alias{solve_poisson}
\title{Resolver la ecuación de Poisson}
\usage{
solve_poisson(forcing, lon, lat)
}
\arguments{
\item{forcing}{vector numérico con el forzante (la vorticidad)}

\item{lon, lat}{vectores numéricos con las ubicaciones del forzante}
}
\value{
Un vector numérico con la función corriente
}
\description{
Resolver la ecuación de Poisson
}
