% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_tesis}
\alias{theme_tesis}
\title{Tema de ggplot2 usado en la tesis}
\usage{
theme_tesis(base_size = 16)
}
\arguments{
\item{base_size}{base font size, given in pts}
}
\description{
Basado en \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}} pero con modificaciones... mínimas (badumtssss)
}
