% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveFlux.R
\name{WaveFlux2}
\alias{WaveFlux2}
\title{Computa actividad de onda}
\usage{
WaveFlux2(psi_z, lon, lat, p = 250, a = 6371000)
}
\arguments{
\item{psi_z}{vector con anomalía zonal de función corriente}

\item{lon, lat}{vectores con ubicaciones}

\item{p}{nivel de presión en hPa}

\item{a}{radio de la Tierra en metros}
}
\description{
Computa actividad de onda
}
