% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstats.R
\name{weighted_correlation}
\alias{weighted_correlation}
\title{Computa correlación pesada}
\usage{
weighted_correlation(x, y, w)
}
\arguments{
\item{x, y}{vectores numéricos a correlacionar}

\item{w}{vector de peso}
}
\description{
Computa correlación pesada
}
