% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-contour-pval.R
\name{geom_contour_pval}
\alias{geom_contour_pval}
\title{Contorno al rededor de p-valores 'significativos'}
\usage{
geom_contour_pval(mapping, p.value = 0.01, size = 0.1, ...)
}
\description{
Contorno al rededor de p-valores 'significativos'
}
